+++ src/memory/RomDatabase.hh
@@ -3,7 +3,6 @@
 
 #include "RomInfo.hh"
 #include "MemBuffer.hh"
-#include "InfoTopic.hh"
 #include "sha1.hh"
 #include <utility>
 #include <vector>
@@ -11,14 +10,13 @@
 namespace openmsx {
 
 class CliComm;
-class GlobalCommandController;
 
 class RomDatabase
 {
 public:
 	using RomDB = std::vector<std::pair<Sha1Sum, RomInfo>>;
 
-	RomDatabase(GlobalCommandController& commandController, CliComm& cliComm);
+	RomDatabase(CliComm& cliComm);
 
 	/** Lookup an entry in the database by sha1sum.
 	 * Returns nullptr when no corresponding entry was found.
@@ -30,14 +28,6 @@ public:
 private:
 	RomDB db;
 	MemBuffer<char> buffer;
-
-	struct SoftwareInfoTopic final : InfoTopic {
-		explicit SoftwareInfoTopic(InfoCommand& openMSXInfoCommand);
-		void execute(array_ref<TclObject> tokens,
-			     TclObject& result) const override;
-		std::string help(const std::vector<std::string>& tokens) const override;
-		void tabCompletion(std::vector<std::string>& tokens) const override;
-	} softwareInfoTopic;
 };
 
 } // namespace openmsx
